refactored skill matrix, added capitalization to modules
