+++ src/components/EditEmpPortfolio.tsx
@@ -21,9 +21,9 @@ import {
 } from "../api/api";
 import "../css/EditEmpPortfolio.css";
 import Matrix from "../interfaces/Matrix";
-import CertificationContainer from "./certifications/CertificationContainer";
-import EducationContainer from "./education/EducationContainer";
-import HonorAwards from "./honors/HonorAward";
+import CertificationContainer from "./Certifications/CertificationContainer";
+import EducationContainer from "./Education/EducationContainer";
+import HonorAwards from "./Honors/HonorAward";
 import IndustryEquivalency from "./IndustryEquivalency";
 import OtherWorkExperience from "./OtherWorkExperience";
 import Project from "./Project";
@@ -36,7 +36,7 @@ import { useHistory } from "react-router-dom";
 
 import { toast } from "react-toastify";
 
-import SkillMatrixContainer from "./SkillMatrixContainer";
+import SkillMatrixContainer from "./SkillMatrix/SkillMatrixContainer";
 
 const EditEmpPortfolio = () => {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
