+++ src/components/education/EducationUpdate.tsx
@@ -3,9 +3,9 @@ import React, { useState, FC, CSSProperties } from 'react'
 import { Button, Modal } from "react-bootstrap";
 import "../../css/Project.css";
 import {url} from "../../api/api";
-import educationValidation, { educationValidationErrors } from "../validation/EducationValidation";
-import styleInvalidElements, { styleInvalidElement } from "../validation/InvalidFormHandling";
-import ValidationMsg from '../validation/ValidationMsg';
+import educationValidation, { educationValidationErrors } from "../Validation/EducationValidation";
+import styleInvalidElements, { styleInvalidElement } from "../Validation/InvalidFormHandling";
+import ValidationMsg from '../Validation/ValidationMsg';
 import EducationData from "../../interfaces/Education";
 import EducationEdit from "./EducationEdit";
